requirements.txt
cycler==0.11.0
Cython==0.29.28
decorator==4.4.2
fonttools==4.33.3
gensim==4.2.0
joblib==1.1.0
karateclub==1.2.3
kiwisolver==1.4.2
matplotlib==3.5.2
metis==0.2a5
networkx==2.8
networkx-metis==1.0
numpy==1.22.3
packaging==21.3
pandas==1.3.5
Pillow==9.1.0
pydot==1.4.2
pygraphviz==1.9
PyGSP==0.5.1
pyparsing==3.0.8
python-dateutil==2.8.2
python-Levenshtein==0.12.2
python-louvain==0.16
pytz==2022.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
smart-open==6.0.0
threadpoolctl==3.1.0
tqdm==4.64.0
